// apps/backend/src/documents/dtos/foreign-registration-form.dto.ts
export class ForeignRegistrationFormDto {
  // TODO add field validations
  // TODO put similar properties into the same key

  addressHomeCountry: string;
  addressKorea: string;
  dateOfBirth: string;
  email: string;
  firstName: string;
  lastName: string;
  mobile: string;
  nationality: string;
  passportExpiryDate: string;
  passportIssueDate: string;
  passportNumber: string;
  phoneNumber: string;
  sex: string;
  telephone: string;
  telephoneNumber: string;
}